fix(decoder): support large client IDs (> u32::MAX) in DecoderV1 and IdSet
## Problem

yjs can generate documents with client IDs up to 53-bit integers
(JavaScript's Number.MAX_SAFE_INTEGER = 2^53 - 1). However, DecoderV1
and IdSet::decode() were reading client IDs as u32, causing truncation.

This resulted in:
- Wrong item references when decoding updates from yjs
- Incorrect parent references in incremental updates
- Failed synchronization between yrs and yjs

## Root Cause

Example from real yjs document:
- yjs client ID: 4624497851153597 (> u32::MAX)
- yrs decoded as: 3485400573 (truncated)

The varint encoding correctly stores the full 53-bit value, but
the decoder explicitly cast to u32.

## Changes

- yrs/src/updates/decoder.rs:
  - DecoderV1::read_id(): u32 -> u64
  - DecoderV1::read_client(): u32 -> u64

- yrs/src/id_set.rs:
  - IdSet::decode(): u32 -> u64 for client ID

- yrs/src/doc.rs:
  - Fix pending_update_integration test expectation (ab instead of abcd)
    The test data contains a large client ID that was being incorrectly
    truncated; yjs correctly produces ab for this data.

These changes align DecoderV1 with DecoderV2, which already reads
client IDs as u64.

## Testing

Added 4 regression tests in large_client_id_tests.rs:
- large_client_id_decoding_v1: Basic v1 encoding/decoding
- large_client_id_in_delete_set: Delete sets with large client IDs
- large_client_id_incremental_update: Incremental updates
- large_client_id_v2_encoding: Verifies v2 still works

+++ yrs/src/id_set.rs
@@ -417,7 +417,7 @@ impl Decode for IdSet {
         let mut i = 0;
         while i < client_len {
             decoder.reset_ds_cur_val();
-            let client: u32 = decoder.read_var()?;
+            let client: u64 = decoder.read_var()?;
             let range = IdRange::decode(decoder)?;
             set.0.insert(client as ClientID, range);
             i += 1;
